Commit Summary.
-Add more trees on the game scene.
-Replace extends Control with extends node.
-Add a game tag line on the menu screen and reposition play button.
-Move playercount into different file instead.
-Move setting isOnline false to accountRoute instead and add a new route for logout where only player is validated, guess account will be deleted.
-Add a send route to gameData script for logout.
-Add a two seconds timeout for clearing the socket username and gameID.
-Remove the player count decrement on logout at websocket, now the only the closed connection will update the playercount.

-TODO: fix the playercount and player logout, sometimes it doesn't work properly.

+++ Server/websocket_game_stuff.js
@@ -1,4 +1,3 @@
-const gameDataModel = require("./gameDataMongooseSchema");
 const accountModel = require("./accountMongooseSchema");
 const playerInfoModel = require("./playerInformationMongooseSchema");
 const broadcastSocket = require("./websocket_broadcast");
@@ -47,19 +46,7 @@ module.exports = (wss)=>{
                 )
                 ws.GameID = parsed_message.Player_GameID;
                 ws.username = parsed_message.Player_username;
-
-                let count = await modifyPlayerCount(1)
-
-                if(count["status"] == "Succeed"){
-                    broadcastSocket(
-                        wss,
-                        {
-                            "Socket_Name": "PlayerCount",
-                            "Count": count["playerCount"]
-                        }
-                    )
-                }
-
+                
                 ws.send(JSON.stringify(
                     {
                         "Socket_Name": socket_name,
@@ -77,22 +64,10 @@ module.exports = (wss)=>{
                         "Player_GameID": parsed_message.GameID
                     }
                 )
-                await deleteGuestPlayer_account(parsed_message.Player_username);
-                
-                ws.GameID = ""
-                ws.username = ""
-
-                let count = await modifyPlayerCount(-1)
-
-                if(count["status"] == "Succeed"){
-                    broadcastSocket(
-                        wss,
-                        {
-                            "Socket_Name": "PlayerCount",
-                            "Count": count["playerCount"]
-                        }
-                    )
-                }
+                setTimeout(() => {
+                    ws.GameID = ""
+                    ws.username = ""
+                }, 2000);
             }
 
             //for player modify profile
@@ -129,19 +104,7 @@ module.exports = (wss)=>{
                         "Player_GameID": ws.GameID
                     }
                 )
-
-                let count = await modifyPlayerCount(-1)
-
-                if(count["status"] == "Succeed"){
-                    broadcastSocket(
-                        wss,
-                        {
-                            "Socket_Name": "PlayerCount",
-                            "Count": count["playerCount"]
-                        }
-                    )
-                }
-
+                console.log(ws.username + " closed the game.")
                 await deleteGuestPlayer_account(ws.username);
             }
         });
@@ -154,20 +117,30 @@ module.exports = (wss)=>{
 
 async function deleteGuestPlayer_account(username) {
     try{
-        let deleteAcc = await accountModel.findOneAndDelete({ username: username, account_type: "Guest" });
+        let findAcc = await accountModel.findOne({ username: username, account_type: "Guest" });
 
-        if(deleteAcc){
+        if(findAcc){
+            console.log("this is a guest account, isOnline: " + findAcc.isOnline)
+            
+            if(findAcc.isOnline){
+                await modifyPlayerCount(-1);
+            }
             const find_player = await playerInfoModel.findOne({ username: username })
 
             if(find_player){
                 await delete_image(find_player.profile_hash);
-                await playerInfoModel.findOneAndDelete({ username: username });
             }
+
+            await playerInfoModel.findOneAndDelete({ username: username });
+            await accountModel.findOneAndDelete({ username: username, account_type: "Guest" });
         }
         else{
-            await accountModel.findOneAndUpdate({ username: username }, { $set: { isOnline: false }, new: true })
+            const findAcc = await accountModel.findOneAndUpdate({ username: username, isOnline: true, account_type: "Player" }, { $set: { isOnline: false }}, { new: true })
+
+            if(!findAcc){
+                console.log("Account already logged out");
+            }
         }
-        
     }
     catch(err){
         console.log(err);
@@ -176,8 +149,7 @@ async function deleteGuestPlayer_account(username) {
 
 async function modifyPlayerCount(count){
     try{
-        let status = "Failed"
-        let playerCount = 0
+        const gameDataModel = require("./gameDataMongooseSchema")
 
         await gameDataModel.findOneAndUpdate(
             {}, 
@@ -191,18 +163,6 @@ async function modifyPlayerCount(count){
             { $max: { playerCount: 0 }},
             { new: true }
         );
-
-        const updatePlayerCount = await gameDataModel.findOne({});
-
-        if(updatePlayerCount){
-            status = "Succeed";
-            playerCount = updatePlayerCount.playerCount
-        }
-
-        return {
-            status: status,
-            playerCount: playerCount
-        }
     }
     catch(err){
         console.log(err)
